UI added statistics

+++ Flask_bootstrap v1 - samo WEBpage/app-emitter.py
@@ -56,8 +56,8 @@ def callback(sender, data, mac_address):
     else:
         statistikaNobena+=1
     ravnotezen=statistikaLeva+statistikaDesna+statistikaNobena
-    leva=statistikaLeva*100/ravnotezen
-    desna=statistikaDesna*100/ravnotezen
+    leva=int(statistikaLeva*100/ravnotezen)
+    desna=int(statistikaDesna*100/ravnotezen)
     number=[number1, number2, leva, desna, (100-leva-desna)] #tale array posljemo preko sock emit na spletno stran
     socketio.emit('newnumber', {'number': number}, namespace='/test')
